Add stella to the player
This will make the player position a list of tuples, where the first
element is the head or the motorcycle, and the other N elements are the
stella. The stella is initialized with None, and will be added with each
iteration of the game loop until the `possition` is `N_STELLA` + 1
elements long. Then, the whole stella moves with the player.