Create `Tron().__collision_happened()` method
This function will check if a collision happened during the game loop.
It will return one of for 4 options:
- `PLAYER_COLLIDED`: If players' head collided into each other.
- `PLAYER_1_COLLIDED`: If player 1 collided into a wall or player 2
stella.
- `PLAYER_2 COLLIDED`: If player 2 collided into a wall or player 1
stella.
- `BOTH_WALLS`: If both players collided into walls.

This commit also creates the `Tron().__init_walls()` method, which
fills the `Tron().__walls` property with the cells where the walls are
located, and fills the `Tron().__board` borders with `WALL`.