Check for suicide and use previous_move
This commit adds the `Player().player_suicided(new_move)` function to
check if the player commited suicide. This means that the player's new
move is the opposite of the player's `previous_move`.

This commit is also the beggining of the implementation of the
game-loop. It starts by picking up the input from the players and
checking if they are valid moves. If not, it will assign the
previous_move as the new_move