Make the game loop more single responsibility
This commit makes the game loop more single responsibility by moving the
different steps of the game loop into their own functions. This makes the
game loop easier to read and understand.

This also fixes the game board, as it was not being updated correctly
because of a pointer issue.